requirements.txt
openai
faster-whisper
pretty_midi
streamlit
sounddevice
numpy
scipy
python-dotenv 